refactor: adjust comments from CR
