+++ example/src/Examples/Dialogs/DialogWithCustomColors.tsx
@@ -7,7 +7,7 @@ import {
   useTheme,
   MD3Colors,
 } from 'react-native-paper';
-import { TextComponent } from './utils';
+import { TextComponent } from './DialogTextComponent';
 const DialogWithCustomColors = ({
   visible,
   close,
@@ -15,27 +15,34 @@ const DialogWithCustomColors = ({
   visible: boolean;
   close: () => void;
 }) => {
-  const theme = useTheme();
+  const { isV3 } = useTheme();
 
   return (
     <Portal>
       <Dialog
         onDismiss={close}
         style={{
-          backgroundColor: theme.isV3
-            ? MD3Colors.primary10
-            : MD2Colors.purple900,
+          backgroundColor: isV3 ? MD3Colors.primary10 : MD2Colors.purple900,
         }}
         visible={visible}
       >
-        <Dialog.Title style={{ color: MD2Colors.white }}>Alert</Dialog.Title>
+        <Dialog.Title
+          style={{ color: isV3 ? MD3Colors.primary95 : MD2Colors.white }}
+        >
+          Alert
+        </Dialog.Title>
         <Dialog.Content>
-          <TextComponent style={{ color: MD2Colors.white }}>
+          <TextComponent
+            style={{ color: isV3 ? MD3Colors.primary95 : MD2Colors.white }}
+          >
             This is a dialog with custom colors
           </TextComponent>
         </Dialog.Content>
         <Dialog.Actions>
-          <Button color={MD2Colors.white} onPress={close}>
+          <Button
+            color={isV3 ? MD3Colors.primary95 : MD2Colors.white}
+            onPress={close}
+          >
             Ok
           </Button>
         </Dialog.Actions>
