+++ example/src/Examples/Dialogs/UndismissableDialog.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { Button, Portal, Dialog, MD2Colors } from 'react-native-paper';
-import { TextComponent } from './utils';
+import { TextComponent } from './DialogTextComponent';
 
 const UndismissableDialog = ({
   visible,
